Fix: Start a flask server on Windows instead of gunicorn server. Fix RequestOrm column type define.
