+++ agentuniverse/agent_serve/web/web_booster.py
@@ -6,79 +6,22 @@
 # @Email   : fanen.lhy@antgroup.com
 # @FileName: web_booster.py
 
-import tomli
-from gunicorn.app.base import BaseApplication
-
-from .web_server import app
-from agentuniverse.base.annotation.singleton import singleton
-
-
-DEFAULT_GUNICORN_CONFIG = {
-    'bind': '127.0.0.1:8000',
-    'workers': 5,
-    'backlog': 2048,
-    'worker_class': 'gthread',
-    'threads': 4,
-    'timeout': 60,
-    'keepalive': 10
-}
-
-
-@singleton
-class GunicornApplication(BaseApplication):
-    """Use gunicorn to wrap the flask web server."""
-    def __init__(self, config_path: str = None):
-        self.options = {}
-        if config_path:
-            self.__load_config_from_file(config_path)
-        else:
-            self.default_config = None
-        self.application = app
-        super().__init__()
-
-    def load_config(self):
-        """Check the config file first, use default config while config file
-        not exist, then overwrite parts which in options."""
-        if not self.default_config:
-            config = DEFAULT_GUNICORN_CONFIG
-        else:
-            config = self.default_config
-        for key, value in config.items():
-            if key in self.cfg.settings and value is not None:
-                self.cfg.set(key.lower(), value)
-
-        # The priority of the passed arguments supersedes that of config file.
-        for key, value in self.options.items():
-            if key in self.cfg.settings and value is not None:
-                self.cfg.set(key.lower(), value)
-
-    def update_config(self, options: dict):
-        self.options = options
-        self.load_config()
-
-    def load(self):
-        return self.application
-
-    def __load_config_from_file(self, config_path: str):
-        """Load gunicorn config file."""
-        try:
-            with open(config_path, 'rb') as f:
-                config = tomli.load(f)["GUNICORN_CONFIG"]
-        except (FileNotFoundError, TypeError):
-            print("can't find gunicorn config file, use default config")
-            return
-        except (tomli.TOMLDecodeError, KeyError):
-            print("gunicorn config file isn't a valid toml, "
-                  "use default config.")
-            return
-
-        self.default_config = {
-            key: value for key, value in config.items()
-        }
+import sys
 
 
 def start_web_server(**kwargs):
-    """Start func of flask server. Accept input arguments to overwrite default
-    gunicorn config."""
-    GunicornApplication().update_config(kwargs)
-    GunicornApplication().run()
+    """Start func of flask server(on windows platform) or gunicorn server(on
+    others). Accept input arguments to overwrite default gunicorn config."""
+    if sys.platform.startswith('win'):
+        from .flask_server import app
+        if 'bind' in kwargs:
+            host, port = kwargs['bind'].split(':')
+            port = int(port)
+        else:
+            port = 8888
+            host = '0.0.0.0'
+        app.run(port=port, host=host, debug=False)
+    else:
+        from .gunicorn_server import GunicornApplication
+        GunicornApplication().update_config(kwargs)
+        GunicornApplication().run()
