+++ pyproject.toml
@@ -23,9 +23,12 @@ gunicorn = "21.2.0"
 chromadb = "0.4.24"
 sphinx = "^7.2.6"
 sphinx-rtd-theme = "^2.0.0"
-aliyun-log-python-sdk = "0.8.8"
+aliyun-log-python-sdk = { version = "0.8.8", optional = true}
 googleapis-common-protos = "^1.63.0"
 myst-parser = "^2.0.0"
+
+[tool.poetry.extras]
+log_ext = ["aliyun-log-python-sdk"]
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.0"
