+++ tests/test_agentuniverse/unit/agent_serve/test_web_booster.py
@@ -1,17 +1,15 @@
-import os
+import pytest
 
 from agentuniverse.agent_serve.web import web_booster
 from agentuniverse.base.agentuniverse import AgentUniverse
 
 
 def test_service():
-    os.environ['OPENAI_API_KEY'] = 'you openai api key'
+    # os.environ['OPENAI_API_KEY'] = 'you openai api key'
     AgentUniverse().start(config_path='../agent/config.toml')
     web_booster.start_web_server(bind="127.0.0.1:8002")
 
 
 if __name__ == "__main__":
-    # os.environ['OPENAI_API_KEY'] = 'you openai api key'
-    AgentUniverse().start(config_path='../agent/config.toml')
-    web_booster.start_web_server(bind="0.0.0.0:8002")
+    pytest.main([__file__, "-s"])
